Plugin refactoring
